Correctif

+++ app/src/main/java/com/example/ihmproject/alertfactory/Alert.java
@@ -1,13 +1,13 @@
 package com.example.ihmproject.alertfactory;
 
 public class Alert {
-    private int longitude;
-    private int latitude;
+    private double longitude;
+    private double latitude;
     private int type; //1==incident, 2==accident
     private String title;
     private String description;
 
-    public Alert(int longitude, int latitude, String title, String description, int type){
+    public Alert(double longitude, double latitude, String title, String description, int type){
         this.longitude = longitude;
         this.latitude = latitude;
         this.description = description;
@@ -15,9 +15,9 @@ public class Alert {
         this.type = type;
     }
 
-    public int getLongitude(){return longitude;}
+    public double getLongitude(){return longitude;}
 
-    public int getLatitude(){return latitude;}
+    public double getLatitude(){return latitude;}
 
     public int getType(){return type;}
