+++ app/src/main/java/com/example/ihmproject/alertfactory/Incident.java
@@ -1,7 +1,7 @@
 package com.example.ihmproject.alertfactory;
 
 public class Incident extends Alert {
-    public Incident(int longitude, int latitude, String title, String description){
+    public Incident(double longitude, double latitude, String title, String description){
         super(longitude, latitude, title, description, 1);
     }
 }
